+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Always default app CR labels so we can retire the legacy `1.0.0` version label.
+- Validate `.spec.catalog` using Catalog CRs instead of AppCatalog CRs.
+
 ## [0.10.1] - 2021-06-16
 
 ### Changed